Improve SM and deploy Subgraph

+++ smart-contracts/contracts/Badge.sol
@@ -1,21 +1,21 @@
 // SPDX-License-Identifier: GPL-3.0
 pragma solidity ^0.8.0;
 
-import "hardhat/console.sol";
+// import "hardhat/console.sol";
 
 contract Badge {
 
   address public owner = msg.sender;
-  uint16 currentId = 0;
+  uint256 currentId = 0;
   struct BadgeInfo{
-    uint16 id;
+    uint256 id;
     address issuer;
     string issuerName;
     address recipient;
     string recipientName;
   }
-  mapping(uint16 => BadgeInfo) public badgesById;
-  event BadgeIssued(uint16 id, string issuerName, string recipientName);
+  mapping(uint256 => BadgeInfo) public badgesById;
+  event BadgeIssued(uint256 id, string issuerName, string recipientName);
 
   function issue(string memory _issuerName, address _recipient, string memory _recipientName) public {
     badgesById[currentId] = BadgeInfo({
